MySchool, esia, weeks, log out, profiles, translation, documentation

+++ app/src/main/java/org/bxkr/octodiary/models/auth/UserAuthenticationForMobileResponse.kt
@@ -9,7 +9,7 @@ data class UserAuthenticationForMobileResponse(
     @SerializedName("mesh_id")
     val meshId: String,
     @SerializedName("obr_id")
-    val obrId: Int,
+    val obrId: Long,
     @SerializedName("request_parameters")
     val requestParameters: RequestParameters,
     @SerializedName("sso_id")
