+++ app/src/main/java/org/bxkr/octodiary/models/mark/Activity.kt
@@ -7,5 +7,5 @@ data class Activity(
     @SerializedName("lesson_topic")
     val lessonTopic: String,
     @SerializedName("schedule_item_id")
-    val scheduleItemId: Int
+    val scheduleItemId: Long
 )
