+++ app/src/main/java/org/bxkr/octodiary/models/mealbalance/ClientId.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ClientId(
     @SerializedName("contractId")
-    val contractId: Int,
+    val contractId: Long,
     @SerializedName("personId")
     val personId: String,
     @SerializedName("staffId")
