+++ app/src/main/java/org/bxkr/octodiary/models/profile/Group.kt
@@ -5,11 +5,11 @@ import com.google.gson.annotations.SerializedName
 
 data class Group(
     @SerializedName("id")
-    val id: Int,
+    val id: Long,
     @SerializedName("is_fake")
     val isFake: Boolean,
     @SerializedName("name")
     val name: String,
     @SerializedName("subject_id")
-    val subjectId: Int
+    val subjectId: Long
 )
